CW and CCW airflow simulation files uploaded
